fix(helper): cast frequency to long long unsigned int

+++ helper.c
@@ -7,7 +7,8 @@ void print_symbol_freq(const SymbolFreq *freq, size_t size) {
     for (size_t i = 0; i < size; i++) {
         if (freq[i].frequency > 0)
             printf("SymbolFreq: %lu %c %llu %f\n", i, freq[i].symbol,
-                   freq[i].frequency, freq[i].probability);
+                   (long long unsigned int)freq[i].frequency,
+                   freq[i].probability);
     }
 }
